requirements.txt
accelerate==1.11.0
anyio==4.11.0
bitsandbytes==0.48.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1762976168352/work/certifi
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==2.1.5
mat73==0.65
matplotlib==3.10.7
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.3
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open_clip_torch==3.2.0
openai-whisper==20250625
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scenedetect==0.6.7.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.21
tokenizers==0.22.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.1
triton==3.2.0
typer-slim==0.20.0
typing_extensions==4.15.0
ultralytics==8.3.229
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
